+++ src/main/java/dev/seasnail1/utilities/events/TotemPopEvent.java
@@ -3,10 +3,10 @@ package dev.seasnail1.utilities.events;
 import net.minecraft.entity.player.PlayerEntity;
 
 public class TotemPopEvent {
-    public int totems;
+    public short totems;
     public PlayerEntity player;
 
-    public TotemPopEvent(int totems, PlayerEntity player) {
+    public TotemPopEvent(short totems, PlayerEntity player) {
         this.totems = totems;
         this.player = player;
     }
